requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
Django==3.2.6
django-admin-honeypot==1.1.0
django-background-task==0.1.8
django-cors-headers==3.8.0
django-filter==21.1
django-otp==1.0.6
djangorestframework==3.12.4
numpy==1.19.5
pandas==1.3.0
regex==2021.7.6
requests==2.25.1
django-admin-honeypot
django-otp==1.0.6
gunicorn==20.1.0
django-heroku==0.3.1
qrcode==6.1
drf-yasg==1.20.0
django-environ==0.8.1
nltk==3.7
textblob==0.17.1
chat-downloader==0.2.0
matplotlib==3.4.2
pandas==1.2.4